backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai>=0.8.0
supabase==2.10.0
python-dotenv==1.0.0
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.12
PyGithub==2.4.0
